fix(providers): per-provider API key input — a Gemini key was overwriting the OpenRouter slot
`ai:setKey` writes `DEFAULT_PROVIDER_ID` ('openrouter') by construction, and Settings offered a
single field labelled "OpenRouter API key". Every provider after the first was therefore
unreachable: a user pasting a Gemini key destroyed their working OpenRouter credential and Gemini
still had none, so the orchestrator skipped it and kept routing to OpenRouter.

- `providers:setKey` / `providers:clearKey` take the provider id. setKey enables that provider and
  leaves priority alone; clearKey leaves it enabled, simply unattemptable.
- `ProviderInfo.keyHint` carries a masked tail (`••••5fcw`) — enough to tell two keys apart when
  re-pasting, never enough to reconstruct one.
- Each row renders its own password field, Save/Remove, and a "Get a key" link.
- Both write paths call `onCredentialChange`, so in-flight requests against the old credential are
  cancelled rather than finishing under a key the user just replaced.

Also fixes a real data-loss bug found by the tests: `apply()` swallowed the failure result, so the
key field cleared its input even when the save failed — for a key shown once on a provider's page,
that is unrecoverable. It now reports success and the field clears only on a save that landed.

Tests cover the credential isolation directly (a key saved for one provider does not touch
another's slot) and the freshness scenario end to end against the real registry and credential
store: after a 429 exhausts the chain, a key added to another provider — or a replaced key on the
same one — is used by the very next run, with no restart and nothing cached. The 429 is a test
double; the OpenRouter free tier is currently exhausted.

+++ packages/core-ai/src/provider/adapters/sse.ts
@@ -35,7 +35,7 @@ export async function* readSseStream(
     for (;;) {
       const result = await reader.read();
       if (result.done) break;
-      buffer += decoder.decode(result.value as Uint8Array, { stream: true });
+      buffer += decoder.decode(result.value, { stream: true });
 
       let newline = buffer.indexOf('\n');
       while (newline >= 0) {
